# vulkan: copy iq4_nl LUT into shared memory