@jeffbolznv 2024-11-20
Interesting, I didn't know AMD/Intel had similar issues.

> Do you know why CUDA and ROCm don't run into this issue? Do they move the static LUT into shared memory automatically?

I'm not sure. I see the LUT uses `static constexpr __device__` and not `__constant__`, so maybe it was just getting written to some device memory that stays in the cache and this has been good enough?