@slaren 2024-11-20
> I'm not sure. I see the LUT uses `static constexpr __device__` and not `__constant__`, so maybe it was just getting written to some device memory that stays in the cache and this has been good enough?

That's probably the case. There is some context here: https://github.com/ggerganov/llama.cpp/pull/4773#discussion_r1441924107