@0cc4m 2024-11-20
Ah yeah, interesting how these kinds of details show up across APIs.